requirements.txt
bcrypt==4.0.1
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
ecdsa==0.19.0
Flask==3.0.3
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
nest-asyncio==1.6.0
packaging==24.1
passlib==1.7.4
plotly==5.24.1
pyasn1==0.6.1
python-dotenv==1.0.1
python-jose==3.3.0
requests==2.32.3
retrying==1.3.4
rsa==4.9
setuptools==75.1.0
six==1.16.0
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
